Added new fbs schema
