+++ src/fbschema/p2pmsg_content.fbs
@@ -1,5 +1,6 @@
 //IDL file for p2p message content schema.
 include "common_schema.fbs";
+include "state_response_schema.fbs";
 namespace fbschema.p2pmsg;
 
 table UserSubmittedMessage {
@@ -69,32 +70,5 @@ table HistoryLedger {
    raw_ledger:[ubyte];
 }
 
-union State_Response{ File_HashMap_Response, Block_Response, Content_Response }
-
-table State_Response_Message{
-    state_response:State_Response;
-}
-
-table Content_Response{
-     full_path: string;
-     content: [State_File_Hash];
-}
-
-table File_HashMap_Response{
-    full_path: string;
-    hash_map:[ubyte];
-}
-
-table Block_Response{
-    full_path: string;
-    block_id:uint32;
-    data: [ubyte];
-}
-
-table State_File_Hash{
-    full_path: string;
-    is_file: bool;
-    file_hash: [ubyte];
-}
 
 root_type Content; //root type for message content
